ci: improve python CI
Configure python version to the one configured by the action. Currently
this is python 3.12, which is available on the `ubuntu-latest` image
without hassles.

Configure dependabot to upgrade the basedpyright and ruff versions used
in the CI. It should ignore the uv files in test/ directory, they might
be big and nobody wants pull requests against them.

Set basedpyright and ruff versions to the ones in use, to be upgraded by
dependabot in PRs.